feat(sensors): huge improvements to all rust microservices with log, refactor, cleanup, dockerfiles and so on...

+++ sensors/consumer/src/models/payload_trait.rs
@@ -2,17 +2,17 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Temperature {
-    pub value: f64,
+    pub value: f32,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Humidity {
-    pub value: f64,
+    pub value: f32,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Light {
-    pub value: f64,
+    pub value: f32,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -21,29 +21,29 @@ pub struct Motion {
 }
 
 pub trait PayloadTrait {
-    fn get_value(&self) -> f64;
+    fn get_value(&self) -> f32;
 }
 
 impl PayloadTrait for Temperature {
-    fn get_value(&self) -> f64 {
+    fn get_value(&self) -> f32 {
         self.value
     }
 }
 
 impl PayloadTrait for Humidity {
-    fn get_value(&self) -> f64 {
+    fn get_value(&self) -> f32 {
         self.value
     }
 }
 
 impl PayloadTrait for Light {
-    fn get_value(&self) -> f64 {
+    fn get_value(&self) -> f32 {
         self.value
     }
 }
 
 impl PayloadTrait for Motion {
-    fn get_value(&self) -> f64 {
+    fn get_value(&self) -> f32 {
         if self.value {
             1.0
         } else {
